requirements.txt
flask==2.0.1
flask-socketio==5.1.1
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
eventlet==0.33.0
plotly==5.3.1
pandas==2.1.1
openai==1.3.5
python-dotenv==1.0.0
python-socketio==5.4.0
python-engineio==4.2.1
textblob==0.15.3
RPi.GPIO==0.7.1 